Total rewrite
I removed the jquery dependancy and added handling for open
weather map's API key, the app now requires that the api key is put into
the URL as `?key=<api>` its a bit annoying but I couldn't think of
a better way. The code is vanilla javascript and currently very messy
but usable.